Fix a number of Windows things, not totally sure about all of them yet.
git-svn-id: svn+ssh://svn.code.sf.net/p/cmusphinx/code/trunk/sphinxbase@8234 94700074-3cef-4d97-a70e-9c8c206c02f5

+++ include/glist.h
@@ -109,7 +109,7 @@ typedef gnode_t *glist_t;	/** Head of a list of gnodes */
 #define gnode_ptr(g)		((g)->data.ptr)
 #define gnode_int32(g)		((g)->data.i)
 #define gnode_uint32(g)		((g)->data.ui)
-#define gnode_float32(g)	((g)->data.fl)
+#define gnode_float32(g)	((float32)(g)->data.fl)
 #define gnode_float64(g)	((g)->data.fl)
 #define gnode_next(g)		((g)->next)
